# clients/typescript/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loom gateway API",
    "description": "Multi-tenant LLM gateway: tenant-scoped conversations and turns over pluggable providers.",
    "contact": {
      "name": "Fieldstate",
      "email": "hello@fieldstate.nz"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/batches": {
      "post": {
        "tags": [
          "batches"
        ],
        "summary": "`POST /v1/batches` — create a batch from a list of items.",
        "description": "Budget and rate limits are enforced up front with the **same** check the\ninteractive turn paths use ([`crate::v1::enforce_limits`]), so a tenant whose\nbudget action is `block` cannot bypass the block by submitting work as a\nbatch. The rate-limit check counts a batch create as a **single** request\n(not one per item); the budget block is the primary guard against a large\nbatch running up spend past a hard cap.",
        "operationId": "create_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Batch created (status = created)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid virtual key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "402": {
            "description": "Budget exceeded (block action)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Capability unsupported or provider not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "virtual_key": []
          }
        ]
      }
    },
...
    "/v1/conversations/{id}": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "`GET /v1/conversations/{id}` — fetch a conversation with a page of its\nhistory, scoped to the caller's tenant.",
        "operationId": "get_conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max messages to return (default 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Messages to skip from the start",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The conversation and a page of its messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "404": {
            "description": "No such conversation for this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "virtual_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "conversations"
        ],
        "summary": "`DELETE /v1/conversations/{id}` — delete a conversation scoped to the\ncaller's tenant.",
        "operationId": "delete_conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "No such conversation for this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "virtual_key": []
          }
        ]
      }
    },
    "/v1/conversations/{id}/turns": {
      "post": {
        "tags": [
          "conversations"
        ],
        "summary": "`POST /v1/conversations/{id}/turns` — append a user turn, run the bound\nprovider, and return (or stream) the assistant turn, persisting both.",
        "operationId": "create_turn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The assistant turn: a JSON `Message` when `stream=false`, or an SSE stream of `TurnEvent` envelopes when `stream=true`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TurnEvent"
                }
              }
            }
          },
          "404": {
            "description": "No such conversation for this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Capability unsupported or provider not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "virtual_key": []
          }
        ]
      }
    },
    "/v1/turns": {
      "post": {
        "tags": [
          "conversations"
        ],
        "summary": "`POST /v1/turns` — run the provider on a fully-inline conversation with **no**\npersistence. Shares [`execute_turn`] with the stateful path for parity.",
        "operationId": "stateless_turn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatelessTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The assistant turn: a JSON `Message` when `stream=false`, or an SSE stream of `TurnEvent` envelopes when `stream=true`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TurnEvent"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Capability unsupported or provider not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "virtual_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchItemInput": {
        "type": "object",
        "description": "One item of a batch create request: the inline stateless-turn shape plus an\noptional caller correlation id.",
        "required": [
          "provider",
          "model",
          "messages"
        ],
        "properties": {
          "custom_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "A caller-facing correlation id, unique within the batch. Defaults to\n`item-{index}` when omitted."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "The full, inline message history to run."
          },
          "model": {
            "type": "string",
            "description": "The model identifier, as the provider expects it."
          },
          "options": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConversationOptions",
                "description": "Request-time provider options."
              }
            ]
          },
          "provider": {
            "type": "string",
            "description": "The provider to run against (e.g. `\"anthropic\"`)."
          },
          "system": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional system prompt."
          },
          "system_cache": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CacheHint",
                "description": "An optional prompt-cache breakpoint on the system prompt."
              }
            ]
          }
        }
      },
...
      "CacheHint": {
        "type": "object",
        "description": "A request to cache the request prefix up to and including the annotated\nelement.\n\nThis is Loom's provider-agnostic spelling of a cache breakpoint (Anthropic's\n`cache_control: { \"type\": \"ephemeral\" }`). A hint is always *ephemeral*;\n[`ttl`](CacheHint::ttl) optionally selects a non-default lifetime.\n\n# Serde representation\n\nSerializes as an object carrying only a non-default `ttl`, so a default hint\nround-trips as `{}` and a hint with an explicit lifetime as\n`{ \"ttl\": \"one_hour\" }`. This keeps the on-the-wire domain form stable and\nprovider-agnostic; the mapping to a provider's native marker lives in that\nprovider's translator.",
        "properties": {
          "ttl": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CacheTtl",
                "description": "The cache lifetime, or `None` for the provider's default (short) TTL."
              }
            ]
          }
        }
      },
...
      "Conversation": {
        "type": "object",
        "description": "A persisted, multi-tenant conversation: its identity, provider binding,\nmessage history, and metadata.\n\nThis is the top-level aggregate of the domain model. It owns the ordered\n[`Message`] history but not the request-time [`ConversationOptions`], which\nare supplied per request by higher layers.\n\n[`ConversationOptions`]: crate::ConversationOptions",
        "required": [
          "id",
          "tenant_id",
          "binding",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "binding": {
            "$ref": "#/components/schemas/ProviderBinding",
            "description": "The provider and model this conversation is bound to."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the conversation was created."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The conversation's unique identifier."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "The ordered message history."
          },
          "metadata": {
            "description": "Free-form, caller-supplied metadata (tags, correlation IDs, …).\n\nDefaults to JSON `null` when unset."
          },
          "system": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional system prompt applied to the conversation."
          },
          "system_cache": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CacheHint",
                "description": "An optional prompt-cache breakpoint on the system prompt.\n\nWhen set (and [`system`](Conversation::system) is present), provider\ntranslators mark the system prefix — which, together with any tools,\nforms the stable head of the request — as a cache breakpoint. This is a\nrequest-render concern rather than durable state: it is **not**\npersisted, so it takes effect only on the in-memory conversation a turn\nis run against (the stateless turn path, or the auto-cache strategy for\npersisted history). It is absent (rather than `null`) when unset."
              }
            ]
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "description": "The tenant that owns this conversation, for multi-tenant isolation and\nattribution."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the conversation was last updated."
          }
        }
      },
      "ConversationOptions": {
        "type": "object",
        "description": "Options that shape how a provider should generate a response.\n\nThe common, cross-provider sampling controls are modelled as typed fields.\nAnything provider-specific — Anthropic's `tool_choice` and `top_p`, cache\ndirectives, and so on — lives in the [`provider_options`] bag, keyed by\nprovider name. This keeps the common path typed while never forcing a\nprovider feature to be expressed as a stringly-typed hack.\n\nThe type is `#[non_exhaustive]` and implements [`Default`], so callers can\nbuild it up field by field:\n\n```\nuse loom_core::ConversationOptions;\n\nlet mut opts = ConversationOptions::new();\nopts.temperature = Some(0.7);\nopts.max_tokens = Some(1024);\nassert_eq!(opts.temperature, Some(0.7));\n```\n\n[`provider_options`]: ConversationOptions::provider_options",
        "properties": {
          "auto_cache": {
            "type": "boolean",
            "description": "Opt-in automatic prompt caching for this request.\n\nWhen `true`, a provider translator deterministically places cache\nbreakpoints for the (typically persisted) conversation — after the\nstable system-plus-tools head, and on the trailing history boundary —\nwithout the caller annotating individual blocks. This is the recommended\npath for persisted conversations, where per-block hints on reconstructed\nhistory would otherwise have to be re-applied every turn. Defaults to\n`false`."
          },
          "cache_negotiation": {
            "$ref": "#/components/schemas/CacheNegotiation",
            "description": "How a provider should treat cache hints on a model that does not support\nprompt caching. Defaults to [`CacheNegotiation::SoftIgnore`] — cache\nhints are advisory."
          },
          "max_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The maximum number of tokens the model may generate in its response.",
            "minimum": 0
          },
          "mcp_servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/McpServerRef"
            },
            "description": "External MCP (Model Context Protocol) servers the provider should\nconnect to on the model's behalf, so the model can call their tools.\n\nEach entry references a server either **by name** — resolved by the\ngateway against a tenant's registered servers, which injects the URL and\n(decrypted) authorization token **server-side** so the secret never\ntransits the client — or **inline** with an explicit\n[`url`](McpServerRef::url) and optional\n[`authorization`](McpServerRef::authorization) (the advanced path). Empty\nwhen none are offered. See [`ServerTool`] for provider-*executed* tools;\nMCP tools are executed by the connected server, brokered by the provider."
          },
          "provider_options": {
            "type": "object",
            "description": "A per-provider bag of native options that Loom does not model as typed\nfields, keyed by provider name (e.g. `\"anthropic\"`).\n\nThis is where provider-specific knobs live — `tool_choice`, `top_p`,\n`top_k`, thinking configuration, cache directives, and so on — as a\nJSON value the provider translator understands. A [`BTreeMap`] is used\nso serialization is deterministic. Empty when unset.",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "server_tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerTool"
            },
            "description": "Definitions of the **provider-executed** (server-side) tools the model\nmay use — web search, code execution, and, via\n[`ServerTool::Raw`], any native tool Loom does not model yet.\n\nUnlike [`tools`](ConversationOptions::tools), the host does not execute\nthese: the provider runs them server-side and returns the results in the\nsame response (as\n[`ServerToolUse`](crate::ContentPart::ServerToolUse) /\n[`ServerToolResult`](crate::ContentPart::ServerToolResult) parts). Empty\nwhen none are offered."
          },
          "stop_sequences": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sequences that, when generated, cause the model to stop. Empty when\nunset."
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Sampling temperature. Provider-defined range; typically `0.0..=1.0`."
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolDefinition"
            },
            "description": "Definitions of the **client-side** tools the model may call. Empty when\nno tools are offered."
          }
        }
      },
      "CreateBatchRequest": {
        "type": "object",
        "description": "Request body for creating a batch.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchItemInput"
            },
            "description": "The batch's items, in submission order."
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A single turn in a conversation.\n\nA message carries an ordered list of [`ContentPart`]s (order is\nsignificant — providers interleave text, tool calls, and reasoning) and,\nfor assistant/provider turns, an optional [`Usage`] snapshot describing what\nthe response cost.",
        "required": [
          "role"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentPart"
            },
            "description": "The message's content, in provider-significant order."
          },
          "raw": {
            "description": "The provider's verbatim native response payload, preserved for audit and\nbyte-equivalent replay.\n\nPopulated by provider translators on the assistant/provider turns they\nsynthesise from a native response, so the exact bytes the provider\nreturned are never lost even where Loom models the response as typed\n[`ContentPart`]s. `None` for messages Loom (or a host application)\nconstructs itself."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "The author of the message."
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Usage",
                "description": "Resource usage reported for this message, when the provider supplied it\n(typically only on assistant/provider responses)."
              }
            ]
          }
        }
      },
      "ProviderBinding": {
        "type": "object",
        "description": "The provider a conversation is bound to, and the model to use.\n\nThe binding is intentionally a pair of plain strings rather than an\nenumeration: providers and their model catalogues evolve independently of\nLoom's release cycle, and Loom must never reject a valid model just because\nit predates a given build.",
        "required": [
          "provider",
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "The model identifier as the provider expects it (e.g.\n`\"claude-opus-4-8\"`)."
          },
          "provider": {
            "type": "string",
            "description": "The provider's name (e.g. `\"anthropic\"`)."
          }
        }
      },
      "Role": {
        "type": "string",
        "description": "Who authored a [`Message`](super::Message).\n\n`#[non_exhaustive]` because providers may introduce further roles (e.g. an\noperator or system channel distinct from the ones modelled here).",
        "enum": [
          "user",
          "assistant",
          "provider"
        ]
      },
...
      "StatelessTurnRequest": {
        "type": "object",
        "description": "Request body for a stateless turn: the whole conversation is supplied inline\nand nothing is persisted.",
        "required": [
          "provider",
          "model",
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "The full, inline message history to run."
          },
          "model": {
            "type": "string",
            "description": "The model identifier, as the provider expects it."
          },
          "options": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConversationOptions",
                "description": "Request-time provider options."
              }
            ]
          },
          "provider": {
            "type": "string",
            "description": "The provider to run against (e.g. `\"anthropic\"`)."
          },
          "stream": {
            "type": "boolean",
            "description": "Whether to stream the assistant turn as Server-Sent Events."
          },
          "system": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional system prompt."
          },
          "system_cache": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CacheHint",
                "description": "An optional prompt-cache breakpoint on the system prompt."
              }
            ]
          }
        }
      },
...
      "TurnEvent": {
        "type": "object",
        "description": "A single streaming event: a normalised envelope plus the verbatim native\nprovider event it was derived from.",
        "required": [
          "kind",
          "raw"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TurnEventKind",
            "description": "The normalised, provider-agnostic view of this event."
          },
          "raw": {
            "description": "The verbatim native provider event JSON this envelope was derived from.\n\nThis is never lossy: whatever the provider sent on the wire is preserved\nhere so clients that need byte-exact fidelity can bypass the normalised\nview entirely."
          }
        }
      },
...
      "TurnRequest": {
        "type": "object",
        "description": "Request body for appending a turn to a stored conversation.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentPart"
            },
            "description": "The user turn's content parts, in provider-significant order."
          },
          "options": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConversationOptions",
                "description": "Request-time provider options (sampling, tools, …)."
              }
            ]
          },
          "stream": {
            "type": "boolean",
            "description": "Whether to stream the assistant turn as Server-Sent Events."
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "admin token"
      },
      "virtual_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "virtual key"
      }
    }
  },
...
}